Create README.md
Hi, everyone! If you are lucky enough to notice my readme, I am so happy. If you want to be a friend with me. That's better. I'm just starting to learn this knowledge. Need a lot of this knowledge of friends to teach me, when my teacher. will you my friend
If you are willing, can you contact me?